routing and not found page added
