+++ package.json
@@ -16,6 +16,7 @@
     "aos": "^2.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scroll": "^1.8.7",
     "remixicon": "^2.5.0",
     "remixicon-react": "^1.0.0",
