+++ src/Pages/style.css
@@ -1,0 +1,45 @@
+.notFound__container {
+  width: 100vw;
+  height: 100vh;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+}
+.notFound__container h2 {
+  font-size: 2.2rem;
+  font-weight: 600;
+}
+.notFound__container p {
+  margin: 25px 0;
+  width: 500px;
+  text-align: center;
+  color: rgba(0, 0, 0, 0.4666666667);
+}
+.notFound__container img {
+  width: 400px;
+  height: 400px;
+}
+.notFound__container .button button {
+  background-color: #007DFE;
+  color: #fff;
+  border-radius: 5px;
+  padding: 10px;
+}
+.notFound__container .button button:hover {
+  -webkit-transform: scale(1.01);
+          transform: scale(1.01);
+}
+.notFound__container .button button:active {
+  -webkit-transform: scale(0.98);
+          transform: scale(0.98);
+}
