+++ src/Pages/style.scss
@@ -1,0 +1,40 @@
+.notFound__container {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+
+  h2 {
+    font-size: 2.2rem;
+    font-weight: 600;
+  }
+
+  p {
+    margin: 25px 0;
+    width: 500px;
+    text-align: center;
+    color: #0007;
+  }
+
+  img {
+    width: 400px;
+    height: 400px;
+  }
+
+  .button {
+    button {
+      background-color: #007DFE;
+      color: #fff;
+      border-radius: 5px;
+      padding: 10px;
+      &:hover {
+        transform: scale(1.01);
+      }
+      &:active {
+        transform: scale(0.98);
+      }
+    }
+  }
+}
